+++ core/src/main/java/io/snabble/sdk/Project.java
@@ -130,22 +130,11 @@ public class Project {
 
         isCheckoutAvailable = JsonUtils.getBooleanOpt(jsonObject, "enableCheckout", true);
 
-        if (jsonObject.has("encodedCodes")) {
-            JsonElement encodedCodes = jsonObject.get("encodedCodes");
+        if (jsonObject.has("qrCodeOffline")) {
+            JsonElement encodedCodes = jsonObject.get("qrCodeOffline");
             if (!encodedCodes.isJsonNull()) {
                 JsonObject object = encodedCodes.getAsJsonObject();
-
-                encodedCodesOptions = new EncodedCodesOptions.Builder(this)
-                        .prefix(JsonUtils.getStringOpt(object, "prefix", ""))
-                        .separator(JsonUtils.getStringOpt(object, "separator", "\n"))
-                        .suffix(JsonUtils.getStringOpt(object, "suffix", ""))
-                        .maxCodes(JsonUtils.getIntOpt(object, "maxCodes", 100))
-                        .finalCode(JsonUtils.getStringOpt(object, "finalCode", ""))
-                        .nextCode(JsonUtils.getStringOpt(object, "nextCode", ""))
-                        .nextCodeWithCheck(JsonUtils.getStringOpt(object, "nextCodeWithCheck", ""))
-                        .maxSizeMm(JsonUtils.getIntOpt(object, "maxSizeMM", -1))
-                        .maxChars(JsonUtils.getIntOpt(object, "maxChars", EncodedCodesOptions.DEFAULT_MAX_CHARS))
-                        .build();
+                encodedCodesOptions = EncodedCodesOptions.fromJsonObject(this, object);
             }
         }